--- a/lib/protocol/SqlString.js
+++ b/lib/protocol/SqlString.js
@@ -72,20 +72,42 @@ SqlString.arrayToList = function arrayToList(array, timeZone) {
 };
 
 SqlString.format = function format(sql, values, stringifyObjects, timeZone) {
-  values = values == null ? [] : [].concat(values);
+  if (values == null) {
+    return sql;
+  }
 
-  var index = 0;
-  return sql.replace(/\?\??/g, function(match) {
-    if (index === values.length) {
-      return match;
-    }
+  if (!(values instanceof Array || Array.isArray(values))) {
+    values = [values];
+  }
 
-    var value = values[index++];
+  var valuesIndex = 0;
+  var chunkIndex = 0;
+  var placeholdersRegex = /\?\??/g;
+  var result = '';
+  var match;
 
-    return match === '??'
-      ? SqlString.escapeId(value)
-      : SqlString.escape(value, stringifyObjects, timeZone);
-  });
+  while (valuesIndex < values.length && (match = placeholdersRegex.exec(sql))) {
+    var value = values[valuesIndex++];
+
+    result += sql.slice(chunkIndex, match.index) + (
+      match[0] === '??'
+        ? SqlString.escapeId(value)
+        : SqlString.escape(value, stringifyObjects, timeZone)
+    );
+
+    chunkIndex = placeholdersRegex.lastIndex;
+  }
+
+  if (chunkIndex === 0) {
+    // Nothing was replaced
+    return sql;
+  }
+
+  if (chunkIndex < sql.length) {
+    result += sql.slice(chunkIndex);
+  }
+
+  return result;
 };
 
 SqlString.dateToString = function dateToString(date, timeZone) {