Improve performance by:
+ Returning early if there are no values to use for formatting
+ Not copying the `values` argument if it is already an Array
+ Parsing the SQL string using the technique from #1390

#### Benchmark Results

| Query | Before (ops/sec) | This PR (ops/sec) | Speedup (`PR/before`) |
|----------|----------------------|------------------------|-------------|
| no values | 1,733,030 | 48,704,838 | 28 |
| 1) simple query | 759,563 | 1,925,109 | 2.5 |
| 2) more values | 582,647 | 1,098,441 | 1.9 |
| 3) a lot of values | 366,547 | 507,246 | 1.4 |
| 4) long query string, few values | 647,084 | 1,772,840 | 2.7 |

See [this gist](https://gist.github.com/nwoltman/7de8311bb69b56bb0002265ad0e5df97) for the code.