@dougwilson 2016-05-31
Thanks, again, @nwoltman! I added you to the `package.json` over the weekend as well :) LGTM besides the two comments I have.