@nwoltman 2016-05-31
Updated, waiting on AppVeyor (@dougwilson do you have the power to cancel those intermediary commits I pushed that got queue up for testing on AppVeyor?)