@dougwilson 2016-05-31
> @dougwilson do you have the power to cancel those intermediary commits I pushed that got queue up for testing on AppVeyor?

Yep! Canceled them.